refactor(nvim-bridge): split pure helpers out of index.ts (#477)
Move parser, formatter, and PiComms prompt-building helpers into a
new nvim-bridge/helpers.ts module so index.ts stays focused on
extension wiring, per the repo convention.

Moved to helpers.ts:
- types EditorState, NvimEvent, CommentRpcRequest
- asObject, toPositiveInteger (pure parsing primitives)
- formatContext
- parseNvimEvent, parseCommentRpcRequest
- buildPiCommsReadPrompt and its private deps (getCurrentCommentContext,
  commentMatchesCurrentContext, formatCommentForRead, formatCommentContext)

Rename nvim-bridge/index.test.ts -> helpers.test.ts to match
slack-bridge convention. The test can now import directly from
./helpers.js and drop the typebox mock / async beforeAll dance.

No behavior change; public extension surface is unchanged aside
from no longer re-exporting these pure helpers from index.ts (they
had no external callers — just the co-located test).

Follow-up from #472.
